caching/requirements.txt
asgiref==3.3.4
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.3
django-extensions==3.1.3
django-filter==2.4.0
django-memcache-status==2.2
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.7.2
idna==2.10
install==1.3.4
mongoengine==0.23.1
oauthlib==3.1.1
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.1.0
pymongo==3.11.4
pyOpenSSL==20.0.1
python-memcached==1.59
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.5
Werkzeug==2.0.1
